Addressing PR Feedback
- updated LanguageCode as an optional parameter to be sent to the server
- removed LanguageCode and Encoding enum usage within the SpeechRecognitionOptions due to variance in between service implementations
- changed confidence and durationInSeconds fields to float type
- moved SpeechOptions and SpeechRecognitionOptions classes into separate files outside of Audio.cs, and moved these along w/ Microphone into dedicated Sound folder for organisation
- added speechRecognitionResult message which was missing from initial commit
