+++ Scripts/Sound.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8dbfa576e63221e4384dd9b90da2a5b5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
