+++ Scripts/Sound/SpeechOptions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9860f438bff59194e9e6226aab6f2578
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
