+++ Scripts/Sound/SpeechRecognitionOptions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 174275855d19d134993e30890652e86e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
